requirements.txt

fastapi
pydantic
python-dotenv
azure-storage-blob
python-docx
PyMuPDF
numpy
scikit-learn
openai